self_run_bot/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.11.2
discord.py==2.1.0
fastapi==0.101.1
httplib2==0.21.0
nltk==3.8.1
numpy==1.21.4
openai==0.26.3
pandas==1.3.4
pydantic==1.10.4
PyGithub==1.57
PyGithub==1.59.1
pymongo==4.3.3
PyMuPDF==1.21.1
pyparsing==3.1.1
python-dotenv==1.0.0
Requests==2.31.0
tiktoken==0.1.2
transformers==4.26.0
uuid=1.30
uvicorn==0.20.0

